requirements.txt
# Audio processing
librosa==0.10.1
pydub==0.25.1
soundfile==0.12.1

# Speech processing
openai-whisper==20231117
SpeechRecognition==3.10.0

# Machine Learning
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.0.3

# Video/Audio downloading
yt-dlp==2023.11.16
requests==2.31.0

# Utilities
python-dotenv==1.0.0
click==8.1.7
tqdm==4.66.1

# Optional: API framework
flask==2.3.3
